client: detect CLOSE_WAIT and try to reconnect
We have been seeing a problem with HAproxy and this gem. Sometimes
a timeout would elapse and HAproxy would close the connection by
sending a FIN. That transitions the state of the TCP connection to
Redis to a half-closed state, CLOSE_WAIT.

This patch detects that condition and tries to reconnect if
possible (ie. no command duplication).

This code requires patches to hiredis and ruby drivers to let
EOFError exceptions bubble up unchanged.